requirements.txt
aiogram==3.1.1
python-dotenv==1.0.0
SQLAlchemy==2.0.21
alembic==1.12.0
asyncpg==0.28.0
redis==5.0.1